refactor: reuse core request timeout policy
