+++ src/daemon/__tests__/session-teardown-budget.test.ts
@@ -2,8 +2,8 @@ import path from 'node:path';
 import { expect, test } from 'vitest';
 import { AGENT_BROWSER_TIMEOUT_MS } from '@agent-device/platform-web';
 import { mkdtempForTestSync } from '../../__tests__/test-utils/tmp-dir.ts';
+import { makeAndroidSession } from '../../__tests__/test-utils/session-factories.ts';
 import { SessionStore } from '../session-store.ts';
-import type { SessionState } from '../types.ts';
 import {
   resolveDaemonSessionTeardownTimeoutMs,
   SCREEN_RECORDING_SESSION_TEARDOWN_BUDGET_MS,
@@ -28,18 +28,7 @@ test('daemon session teardown budget extends for an active durable recording', (
 
 test('daemon session teardown budget extends for an open web session', () => {
   const webSession = makeWebSession('budget-web-session');
-  const androidSession: SessionState = {
-    name: 'budget-android-session',
-    device: {
-      platform: 'android',
-      id: 'emulator-5554',
-      name: 'Pixel',
-      kind: 'emulator',
-      booted: true,
-    },
-    createdAt: Date.now(),
-    actions: [],
-  };
+  const androidSession = makeAndroidSession('budget-android-session');
 
   expect(
     resolveDaemonSessionTeardownTimeoutMs(webSession) -
@@ -47,16 +36,6 @@ test('daemon session teardown budget extends for an open web session', () => {
   ).toBe(WEB_BROWSER_SESSION_TEARDOWN_BUDGET_MS);
 });
 
-// Pins the "mirrors AGENT_BROWSER_TIMEOUT_MS" comment on WEB_BROWSER_SESSION_TEARDOWN_BUDGET_MS
-// as a checked invariant rather than a claim nothing enforces: if agent-browser-provider.ts's own
-// per-call timeout changes, this budget must move with it in the same PR.
 test('the web-close teardown budget stays pinned to one agent-browser CLI call timeout', () => {
   expect(WEB_BROWSER_SESSION_TEARDOWN_BUDGET_MS).toBe(AGENT_BROWSER_TIMEOUT_MS);
-});
-
-test('command admission reserves recording teardown before a session exists', () => {
-  expect(
-    resolveDaemonSessionTeardownTimeoutMs(undefined, true) -
-      resolveDaemonSessionTeardownTimeoutMs(),
-  ).toBe(SCREEN_RECORDING_SESSION_TEARDOWN_BUDGET_MS);
 });
